Deltamethod
The first R-code is the code that was used for the simulationstudies.
With this code we calculated the power and Type I error rate and
compared it with the results from the Chi²-difference tests. The
SRM-means were taken into account.
The second code is an R-code that calculates the SRM variances with this
deltamethod. It was one of the first codes so I also compared the result
of this method with the result of the chi²-difference test in terms of
their p-values. They seemed to be identical. Therefor the
simulationstudies (code that is described above) where performed.