v4.0:
		\item Removed \code{setZero} option from \code{fSRM}. Replaced it
with the option \code{noNegVar}. If this is TRUE, all variances are
constrained to be > 0. (Older publications contrained negative
variances to be zero `==0'. This, however, leads to a distortion of the
df).
		\item Added \code{rolesEqual} option to \code{fSRM}: Do roles matter
at all? Compare a model with free roles (\code{m1 <- fSRM(...,
means=TRUE, rolesEqual = FALSE)}) and a model with equal roles
(\code{m2 <- fSRM(..., means=TRUE, rolesEqual=TRUE)}) using
\code{anova(m1$fit, m2$fit)}.
		\item If missing values are present, by default \code{missing =
"fiml"} is set, and a warning is issued that this is only valid if the
data are missing completely at random (MCAR) or missing at random
(MAR). Use \code{missing = "listwise"} to exclude families with missing
values.
		\item Moved all dependencies to \code{Imports} so that no manual
installation of packages is necessary for unexperienced users.
		\item By default, one-sided p values and CIs are printed for
variances.
		\item Added plotting methods for fSRM objects.